requirements.txt
click==3.3
python-cinderclient==1.1.1
python-glanceclient==0.16.1
python-keystoneclient==1.2.0
python-neutronclient==2.3.11
python-novaclient==2.22.0
python-openstackclient==1.0.2
python-openstacksdk==0.4.1
